Add barebones Cloud SDK Debug Target
