+++ plugins/com.google.cloud.tools.eclipse.appengine.localserver/src/com/google/cloud/tools/eclipse/appengine/localserver/messages.properties
@@ -20,6 +20,8 @@ SERVER_ALREADY_RUNNING=Server is already running
 SERVER_ALREADY_RUNNING_IN_MODE=Server is already running in "{0}" mode
 UNABLE_TO_LAUNCH=Unable to launch App Engine server
 
+cloudsdk.server.description=Google Cloud SDK Dev App Server
+cloudsdk.server.description.version=Google Cloud SDK Dev App Server [{0}]
 cloudsdk.not.configured=Could not run project because the Cloud SDK was not found. \
  Check Preferences > Google Cloud Tools > SDK location and try again.
 cloudsdk.out.of.date=Could not run project because the installed Cloud SDK is too old. \
